handbook-app/requirements.txt
gradio
pypdf2
python-dotenv
openai
pdfplumber
chromadb
langchain
langchain-community